Split focused command tests by package boundary
