+++ cmd/drift/drift_test.go
@@ -1,4 +1,4 @@
-package drift
+package drift_test
 
 import (
 	"bytes"
@@ -6,14 +6,14 @@ import (
 
 	qt "github.com/frankban/quicktest"
 
+	"github.com/stokaro/ptah/cmd/drift"
 	"github.com/stokaro/ptah/cmd/internal/exitcode"
-	"github.com/stokaro/ptah/migration/safety"
 )
 
 func TestNewDriftCommand_Creation(t *testing.T) {
 	c := qt.New(t)
 
-	cmd := NewDriftCommand()
+	cmd := drift.NewDriftCommand()
 
 	c.Assert(cmd, qt.IsNotNil)
 	c.Assert(cmd.Use, qt.Equals, "drift")
@@ -23,7 +23,7 @@ func TestNewDriftCommand_Creation(t *testing.T) {
 func TestRunDrift_MissingDatabaseURLReturnsCode2(t *testing.T) {
 	c := qt.New(t)
 
-	cmd := NewDriftCommand()
+	cmd := drift.NewDriftCommand()
 	var stderr bytes.Buffer
 	cmd.SetErr(&stderr)
 	cmd.SetOut(&bytes.Buffer{})
@@ -34,64 +34,4 @@ func TestRunDrift_MissingDatabaseURLReturnsCode2(t *testing.T) {
 	c.Assert(err, qt.IsNotNil)
 	c.Assert(exitcode.Code(err, 0), qt.Equals, 2)
 	c.Assert(stderr.String(), qt.Contains, "database URL is required")
-}
-
-func TestParseIgnoredTables(t *testing.T) {
-	c := qt.New(t)
-
-	tables, err := parseIgnoredTables([]string{"tables=audit_log,sessions", "tables= audit_log , events "})
-
-	c.Assert(err, qt.IsNil)
-	c.Assert(tables, qt.DeepEquals, []string{"audit_log", "events", "sessions"})
-}
-
-func TestParseIgnoredTablesRejectsUnknownScope(t *testing.T) {
-	c := qt.New(t)
-
-	_, err := parseIgnoredTables([]string{"views=audit_view"})
-
-	c.Assert(err, qt.ErrorMatches, `invalid --ignore value "views=audit_view": expected tables=name\[,name\.\.\.\]`)
-}
-
-func TestShouldFailDrift(t *testing.T) {
-	c := qt.New(t)
-
-	c.Assert(shouldFailDrift(safety.Warning, severityAll), qt.IsTrue)
-	c.Assert(shouldFailDrift(safety.Warning, severityDestructive), qt.IsFalse)
-	c.Assert(shouldFailDrift(safety.Destructive, severityDestructive), qt.IsTrue)
-}
-
-func TestWriteGitHubActionsReport(t *testing.T) {
-	c := qt.New(t)
-
-	var buf bytes.Buffer
-	err := writeGitHubActionsReport(&buf, driftReport{
-		Drift:            true,
-		Failed:           true,
-		FailureThreshold: severityDestructive,
-		HighestSeverity:  safety.Destructive,
-		Findings: []safety.Finding{
-			{Category: "tables_removed", Count: 1, Severity: safety.Destructive},
-		},
-	})
-
-	c.Assert(err, qt.IsNil)
-	c.Assert(buf.String(), qt.Contains, "::error title=Ptah schema drift::")
-	c.Assert(buf.String(), qt.Contains, "tables_removed: 1")
-}
-
-func TestWriteJSONReport(t *testing.T) {
-	c := qt.New(t)
-
-	var buf bytes.Buffer
-	err := writeReport(&buf, formatJSON, driftReport{
-		Drift:            true,
-		Failed:           false,
-		FailureThreshold: severityDestructive,
-		HighestSeverity:  safety.Warning,
-	})
-
-	c.Assert(err, qt.IsNil)
-	c.Assert(buf.String(), qt.Contains, `"drift": true`)
-	c.Assert(buf.String(), qt.Contains, `"highest_severity": "warning"`)
 }
